Fix all sign comparison warnings
- Fixed 15 sign comparison warnings across multiple files
- CvCultureClasses.cpp: Fixed container size comparisons with static_cast<int>
- CvDealAI.cpp: Fixed GetDuration() vs -1 comparison with static_cast<uint>(-1)
- CvPlayer.cpp: Fixed PlayerOptionTypes enum vs uint comparisons and container size comparisons
- CvTradeClasses.cpp: Fixed trade unit location index comparisons with static_cast<uint>
- CvSiteEvaluationClasses.cpp: Fixed ignorePlots size vs numPlots comparison
- CvTacticalAnalysisMap.cpp: Fixed plot zone ID size vs numPlots comparison
- CvPlotManager.cpp: Fixed layer ID vs -1 comparison with static_cast<uint>(-1)
- CvEspionageClasses.cpp: Fixed spy index vs -1 comparison with static_cast<uint>(-1)
- CvDllDatabaseUtility.cpp: Fixed table ID comparison with static_cast<int>
- CvGame.cpp: Fixed Windows API BOOL vs DWORD comparison

Total warnings reduced from 72 to 57 (21% reduction)
Sign comparison warnings: ELIMINATED (15 → 0)

+++ CvGameCoreDLL_Expansion2/CvAIOperation.cpp
@@ -430,7 +430,7 @@ bool CvAIOperation::RecruitUnit(CvUnit* pUnit)
 		//stupid but true, missing units are maintained on operation level
 		for (std::deque<OperationSlot>::iterator it = m_viListOfUnitsWeStillNeedToBuild.begin(); it != m_viListOfUnitsWeStillNeedToBuild.end(); ++it)
 		{
-			if (it->m_iSlotID == freeSlotInfo[iIndex].first)
+			if (it->m_iSlotID == static_cast<int>(freeSlotInfo[iIndex].first))
 			{
 				m_viListOfUnitsWeStillNeedToBuild.erase(it);
 				break;
